backend/requirements.txt
Django>=5.1,<6.0
channels[daphne]>=4.1.0
channels-redis>=4.2.0
psycopg[binary]>=3.2.0
google-genai>=0.2.0
celery>=5.4.0
redis>=5.0.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
python-decouple>=3.8
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-django>=4.5.0
pytest-cov>=4.1.0